Reorganize views and slices
