+++ include/noarr/structures_extended.hpp
@@ -4,7 +4,8 @@
  */
 
 #include "structures/layouts.hpp"
-#include "structures/view.hpp"
+#include "structures/views.hpp"
+#include "structures/slice.hpp"
 #include "structures/iterator.hpp"
 #include "structures/funcs.hpp"
 #include "structures/wrapper.hpp"
